QuantLib.Instrument.Bond

Documentation

type Bond = GenBond CBond

type FixedRateBond = GenBond CFixedRateBond

type ConvertibleBond = GenBond CConvertibleBond

type CallableBond = GenBond CCallableBond

type CPIBond = GenBond CCPIBond

asBond :: GenBond b -> IO Bond

data BondPriceType

Constructors

Dirty

Clean

data CPIInterpolationType

Constructors

CPIFlat

CPILinear

bond

Arguments

:: Word

-> Calendar

-> Maybe Day
issueDate

-> GenLeg l
coupons

-> IO Bond

constructor for amortizing or non-amortizing bonds. Redemptions and maturity are calculated from the coupon data, if available. Therefore, redemptions must not be included in the passed cash flows.

bond'

Arguments

:: Word
settlementDays

-> Calendar

-> Double
faceAmount

-> Maybe Day
maturityDate

-> Maybe Day
issueDate

-> GenLeg l
cashFlows

-> IO Bond

old constructor for non amortizing bonds. Warning The last passed cash flow must be the bond redemption. No other cash flow can have a date later than the redemption date.

fixedRateBond

Arguments

:: Word
settlementDays

-> Double
faceAmount

-> Schedule
schedule

-> [Double]
coupons

-> DayCounter
accrualDayCounter

-> BusinessDayConvention
paymentConvention

-> Double
redemption

-> Maybe Day
issueDate

-> Calendar
paymentCalendar

-> (Int, TimeUnit)
exCouponPeriod

-> Calendar
exCouponCalendar

-> BusinessDayConvention
exCouponConvention

-> Bool
exCouponEndOfMonth

-> DayCounter
firstPeriodDayCounter

-> IO FixedRateBond

generic compounding and frequency InterestRate coupons

zeroCouponBond

Arguments

:: Word
settlementDays

-> Calendar

-> Double
faceAmount

-> Day
maturityDate

-> BusinessDayConvention

-> Double
redemption

-> Maybe Day
issueDate

-> IO Bond

zero-coupon bond

floatingRateBond

Arguments

:: Word
settlementDays

-> Double
faceAmount

-> Schedule
schedule

-> GenIborIndex ibor

-> DayCounter
accrualDayCounter

-> BusinessDayConvention

-> Word
fixingDays

-> [Double]
gearings

-> [Double]
spreads

-> [Double]
caps

-> [Double]
floors

-> Bool
inArrears

-> Double
redemption

-> Maybe Day
issueDate

-> (Int, TimeUnit)
exCouponPeriod

-> Calendar
exCouponCalendar

-> BusinessDayConvention
exCouponConvention

-> Bool
exCouponEndOfMonth

-> BusinessDayConvention
fixingConvention

-> IO Bond

floating-rate bond (possibly capped and/or floored)

cmsRateBond

Arguments

:: Word
settlementDays

-> Double
faceAmount

-> Schedule
schedule

-> GenSwapIndex sidx

-> DayCounter
paymentDayCounter

-> BusinessDayConvention
paymentConvention

-> Word
fixingDays

-> [Double]
gearings

-> [Double]
spreads

-> [Double]
caps

-> [Double]
floors

-> Bool
inArrears

-> Double
redemption

-> Maybe Day
issueDate

-> IO Bond

CMS-rate bond

cpiBond

Arguments

:: Word
settlementDays

-> Double
faceAmount

-> Double
baseCPI

-> (Word, TimeUnit)
observationLag

-> ZeroInflationIndex

-> CPIInterpolationType
observationInterpolation

-> Schedule

-> [Double]
coupons

-> DayCounter
accrualDayCounter

-> BusinessDayConvention
paymentConvention

-> Maybe Day
issueDate

-> Calendar
paymentCalendar

-> (Int, TimeUnit)
exCouponPeriod

-> Calendar
exCouponCalendar

-> BusinessDayConvention
exCouponConvention

-> Bool
exCouponEndOfMonth

-> IO CPIBond

An inflation-linked bond whose redemption and coupons scale with a ZeroInflationIndex fixing relative to baseCPI.

amortizingFixedRateBond

Arguments

:: Word
settlementDays

-> [Double]
notionals

-> Schedule
schedule

-> [Double]
coupons

-> DayCounter
accrualDayCounter

-> BusinessDayConvention
paymentConvention

-> Maybe Day
issueDate

-> (Int, TimeUnit)
exCouponPeriod

-> Calendar
exCouponCalendar

-> BusinessDayConvention
exCouponConvention

-> Bool
exCouponEndOfMonth

-> [Double]
redemptions

-> Int
paymentLag

-> IO Bond

amortizing fixed-rate bond: like fixedRateBond but with a per-period notional schedule instead of a single face amount (see sinkingSchedule/sinkingNotionals for building one).

amortizingCmsRateBond

Arguments

:: Word
settlementDays

-> [Double]
notionals

-> Schedule
schedule

-> GenSwapIndex sidx

-> DayCounter
paymentDayCounter

-> BusinessDayConvention
paymentConvention

-> Word
fixingDays

-> [Double]
gearings

-> [Double]
spreads

-> [Double]
caps

-> [Double]
floors

-> Bool
inArrears

-> Maybe Day
issueDate

-> [Double]
redemptions

-> IO Bond

amortizing CMS-rate bond (possibly capped and/or floored) with a per-period notional schedule instead of a single face amount, and a per-period redemption schedule instead of a single redemption value.

data AmortizingFloatingRateBondOpts

Constructors

AmortizingFloatingRateBondOpts

Fields

afrbPaymentConvention :: BusinessDayConvention

afrbFixingDays :: Maybe Word

afrbGearings :: [Double]

afrbSpreads :: [Double]

afrbCaps :: [Double]

afrbFloors :: [Double]

afrbInArrears :: Bool

afrbIssueDate :: Maybe Day

afrbExCouponPeriod :: (Int, TimeUnit)

afrbExCouponCalendar :: Maybe Calendar

afrbExCouponConvention :: BusinessDayConvention

afrbExCouponEndOfMonth :: Bool

afrbRedemptions :: [Double]

afrbPaymentLag :: Int

defaultAmortizingFloatingRateBondOpts :: AmortizingFloatingRateBondOpts

amortizingFloatingRateBond :: Word -> [Double] -> Schedule -> GenIborIndex ibor -> DayCounter -> AmortizingFloatingRateBondOpts -> IO Bond

amortizing floating-rate bond (possibly capped and/or floored) with a per-period notional schedule instead of a single face amount; see AmortizingFloatingRateBondOpts for the trailing optional parameters (default via defaultAmortizingFloatingRateBondOpts, override with record-update syntax).

sinkingSchedule

Arguments

:: Day
startDate

-> (Int, TimeUnit)
bondLength

-> Frequency

-> Calendar
paymentCalendar

-> IO Schedule

returns a schedule for French amortization

sinkingNotionals

Arguments

:: (Int, TimeUnit)
bondLength

-> Frequency

-> Double
couponRate

-> Double
initialNotional

-> IO [Double]

returns a sequence of notionals for French amortization

maturityDate :: GenBond b -> Maybe Day

Returns the maturity date of the bond

yield

Arguments

:: GenBond b

-> DayCounter

-> Compounding

-> Frequency

-> Double
accuracy

-> Word
maxEvaluations

-> (Double, BondPriceType)
guess, priceType

-> IO Double

theoretical bond yield

accruedAmount :: GenBond b -> Day -> IO Double

accrued amount at a given date

cleanPriceFromYield :: GenBond b -> Double -> DayCounter -> Compounding -> Frequency -> Day -> IO Double

clean price given a yield and settlement date

dirtyPriceFromYield :: GenBond b -> Double -> DayCounter -> Compounding -> Frequency -> Day -> IO Double

dirty price given a yield and settlement date

nextCashFlowDate :: GenBond b -> Day -> IO (Maybe Day)

date of the next cash flow after the given (or default settlement) date

nextCouponRate :: GenBond b -> Day -> IO Double

Expected next coupon: depending on (the bond and) the given date the coupon can be historic, deterministic or expected in a stochastic sense. When the bond settlement date is used the coupon is the already-fixed not-yet-paid one.The current bond settlement is used if no date is given.

notional :: GenBond b -> Day -> IO Double

bond notional outstanding at the given date

previousCashFlowDate :: GenBond b -> Day -> IO (Maybe Day)

date of the cash flow immediately before the given (or default settlement) date

previousCouponRate :: GenBond b -> Day -> IO Double

Previous coupon already paid at a given date. Expected previous coupon: depending on (the bond and) the given date the coupon can be historic, deterministic or expected in a stochastic sense. When the bond settlement date is used the coupon is the last paid one.The current bond settlement is used if no date is given.

settlementValueFromCleanPrice :: GenBond b -> Double -> IO Double

settlement value as a function of the clean price The default bond settlement date is used for calculation.

settlementValue :: GenBond b -> IO Double

theoretical settlement value The default bond settlement date is used for calculation.

yieldFromPrice

Arguments

:: GenBond b

-> (Double, BondPriceType)

-> DayCounter

-> Compounding

-> Frequency

-> Day

-> Double
accuracy

-> Word
maxEvaluations

-> IO Double

yield given a (clean) price and settlement date

isTradable :: GenBond b -> Day -> IO Bool

whether the bond can be traded (i.e. still has a positive notional) at the given date

notionals :: GenBond b -> IO [Double]

notionals for each period of the bond's amortization schedule

cashFlows :: GenBond b -> IO Leg

returns all the cashflows, including the redemptions.

redemptions :: GenBond b -> IO Leg

returns just the redemption flows (not interest payments)

settlementDate :: GenBond b -> Day -> IO Day

settlement date computed from the given date (or today's date if none is given)

startDate :: GenBond b -> IO Day

date the bond starts accruing

accrualDays :: GenBond b -> Day -> IO Int

number of days in the current accrual period up to the given (or default settlement) date

accrualEndDate :: GenBond b -> Day -> IO (Maybe Day)

end date of the accrual period containing the given (or default settlement) date

accrualPeriod :: GenBond b -> Day -> IO Double

length in time of the accrual period containing the given (or default settlement) date

accrualStartDate :: GenBond b -> Day -> IO (Maybe Day)

start date of the accrual period containing the given (or default settlement) date

accruedDays :: GenBond b -> Day -> IO Int

number of days accrued up to the given (or default settlement) date

accruedPeriod :: GenBond b -> Day -> IO Double

length in time accrued up to the given (or default settlement) date

atmRate :: GenBond b -> GenYieldTermStructure y -> Day -> (Double, BondPriceType) -> IO Double

the bond's yield to maturity given a market price and discount curve

basisPointValue' :: GenBond b -> InterestRate -> Day -> IO Double

basis-point value given an InterestRate yield

basisPointValue :: GenBond b -> Double -> DayCounter -> Compounding -> Frequency -> Day -> IO Double

basis-point value given a flat yield, day counter, compounding and frequency

bpsFromYield :: GenBond b -> Double -> DayCounter -> Compounding -> Frequency -> Day -> IO Double

bps (Basis Point Sensitivity) given a flat yield, day counter, compounding and frequency

bpsFromYield' :: GenBond b -> InterestRate -> Day -> IO Double

bps (Basis Point Sensitivity) given an InterestRate yield

bps :: GenBond b -> GenYieldTermStructure y -> Day -> IO Double

bps (Basis Point Sensitivity) given a discount curve

cleanPrice :: GenBond b -> GenYieldTermStructure y -> Day -> IO Double

clean price given a discount curve and settlement date

cleanPrice'

Arguments

:: GenBond b

-> GenYieldTermStructure y
discount

-> Double
zSpread

-> Compounding

-> Frequency

-> Day
settlementDate

-> IO Double

clean price given a discount curve, a Z-spread over it, compounding and frequency

cleanPriceFromYield' :: GenBond b -> InterestRate -> Day -> IO Double

clean price given an InterestRate yield

convexity'

Arguments

:: GenBond b

-> InterestRate
yield

-> Day
settlementDate

-> IO Double

convexity given an InterestRate yield

convexity

Arguments

:: GenBond b

-> Double
yield

-> DayCounter

-> Compounding

-> Frequency

-> Day
settlementDate

-> IO Double

convexity given a flat yield, day counter, compounding and frequency

duration'

Arguments

:: GenBond b

-> InterestRate
yield

-> DurationType

-> Day

-> IO Double

duration given an InterestRate yield and duration type

duration

Arguments

:: GenBond b

-> Double
yield

-> DayCounter

-> Compounding

-> Frequency

-> DurationType

-> Day
settlementDate

-> IO Double

duration given a flat yield, day counter, compounding, frequency and duration type

nextCashFlowAmount :: GenBond b -> Day -> IO Double

amount of the cash flow immediately after the given (or default settlement) date

previousCashFlowAmount :: GenBond b -> Day -> IO Double

amount of the cash flow immediately before the given (or default settlement) date

referencePeriodEnd :: GenBond b -> Day -> IO (Maybe Day)

end date of the reference period containing the given (or default settlement) date

referencePeriodStart :: GenBond b -> Day -> IO (Maybe Day)

start date of the reference period containing the given (or default settlement) date

yieldFromPrice'

Arguments

:: GenBond b

-> (Double, BondPriceType)

-> DayCounter

-> Compounding

-> Frequency

-> Day
settlementDate

-> Double

-> Word
maxIterations

-> Double
guess

-> IO Double

yield given a (clean) price and settlement date, solved to the given accuracy

yieldValueBasisPoint'

Arguments

:: GenBond b

-> InterestRate
yield

-> Day

-> IO Double

yield value of a basis point given an InterestRate yield

yieldValueBasisPoint

Arguments

:: GenBond b

-> Double
yield

-> DayCounter

-> Compounding

-> Frequency

-> Day

-> IO Double

yield value of a basis point given a flat yield, day counter, compounding and frequency

zSpread

Arguments

:: GenBond b

-> (Double, BondPriceType)

-> GenYieldTermStructure y

-> Compounding

-> Frequency

-> Day
settlementDate

-> Double
accuracy

-> Word
maxIterations

-> Double
guess

-> IO Double

Z-spread over a discount curve implied by a (clean) price, solved to the given accuracy

currentCleanPrice :: GenBond b -> IO Double

theoretical clean price for the current evaluation date and term structure

currentDirtyPrice :: GenBond b -> IO Double

theoretical dirty price The default bond settlement is used for calculation. Warning the theoretical price calculated from a flat term structure might differ slightly from the price calculated from the corresponding yield by means of the other overload of this function. If the price from a constant yield is desired, it is advisable to use such other overload.

callableFixedRateBond

Arguments

:: Word
settlementDays

-> Double
faceAmount

-> Schedule

-> [Double]
coupons

-> DayCounter

-> BusinessDayConvention

-> Double
redemption

-> Maybe Day
issueDate

-> [Callability]

-> (Int, TimeUnit)
exCouponPeriod

-> Calendar
exCouponCalendar

-> BusinessDayConvention
exCouponConvention

-> Bool
exCouponEndOfMonth

-> IO CallableBond

fixed-rate bond with an embedded call/put schedule

callableZeroCouponBond

Arguments

:: Word
settlementDays

-> Double
faceAmount

-> Calendar

-> Day
maturityDate

-> DayCounter

-> BusinessDayConvention

-> Double
redemption

-> Maybe Day
issueDate

-> [Callability]

-> IO CallableBond

zero-coupon bond with an embedded call/put schedule

convertibleFixedCouponBond

Arguments

:: Exercise

-> Double
conversionRatio

-> [Callability]

-> Day
issueDate

-> Word
settlementDays

-> [Double]
coupons

-> DayCounter

-> Schedule

-> Double
redemption

-> (Int, TimeUnit)
exCouponPeriod

-> Calendar
exCouponCalendar

-> BusinessDayConvention
exCouponConvention

-> Bool
exCouponEndOfMonth

-> IO ConvertibleBond

convertible bond with a fixed-rate coupon leg

convertibleFloatingRateBond

Arguments

:: Exercise

-> Double
conversionRatio

-> [Callability]

-> Day
issueDate

-> Word
settlementDays

-> GenIborIndex ibor

-> Word
fixingDays

-> [Double]
spreads

-> DayCounter

-> Schedule

-> Double
redemption

-> (Int, TimeUnit)
exCouponPeriod

-> Calendar
exCouponCalendar

-> BusinessDayConvention
exCouponConvention

-> Bool
exCouponEndOfMonth

-> IO ConvertibleBond

convertible bond with a floating-rate coupon leg

convertibleZeroCouponBond

Arguments

:: Exercise

-> Double
conversionRatio

-> [Callability]

-> Day
issueDate

-> Word
settlementDays

-> DayCounter

-> Schedule

-> Double

-> IO ConvertibleBond

convertible zero-coupon bond