QuantLib.Instrument.Forward

Documentation

type Forward = GenForward CForward

asForward :: GenForward f -> IO Forward

type ForwardRateAgreement = GenInstrument CForwardRateAgreement

type BondForward = GenForward CBondForward

type FxForward = GenInstrument CFxForward

forwardRateAgreement

Arguments

:: GenIborIndex ibor

-> Day
valueDate

-> Day
maturityDate

-> PositionType

-> Double
strikeForwardRate

-> Double
notionalAmount

-> Maybe (GenYieldTermStructure y)
discountCurve

-> IO ForwardRateAgreement

FRA with a par-rate approximation: the forward rate is forecast from valueDate to maturityDate by the index's forecast curve (useIndexedCoupon=false).

bondForward

Arguments

:: Day
valueDate

-> Day
maturityDate

-> PositionType

-> Double
strike

-> Word
settlementDays

-> DayCounter

-> Calendar

-> BusinessDayConvention

-> GenBond b

-> Maybe (GenYieldTermStructure y1)
discountCurve

-> Maybe (GenYieldTermStructure y2)
incomeDiscountCurve

-> IO BondForward

If strike is given in the constructor, can calculate the NPV of the contract via NPV().If strike/forward price is desired, it can be obtained via forwardPrice(). In this case, the strike variable in the constructor is irrelevant and will be ignored.

fxForward

Arguments

:: Double
sourceNominal

-> Currency
sourceCurrency

-> Double
targetNominal

-> Currency
targetCurrency

-> Day
maturityDate

-> Bool
paySourceCurrency

-> Word
settlementDays

-> Calendar
paymentCalendar

-> IO FxForward

FX forward using nominal amounts in both currencies.

fxForward'

Arguments

:: Double
sourceNominal

-> Currency
sourceCurrency

-> Currency
targetCurrency

-> Double
forwardRate

-> Day
maturityDate

-> Bool
paySourceCurrency

-> Word
settlementDays

-> Calendar
paymentCalendar

-> IO FxForward

FX forward using a source nominal amount and a contracted forward rate (target/source).

cleanForwardPrice :: BondForward -> IO Double

(dirty) forward bond price minus accrued on bond at delivery

forwardPrice :: BondForward -> IO Double

(dirty) forward bond price

forwardValue :: GenForward f -> IO Double

forward valueprice of underlying, discounting incomedividends if this is a bond forward price, is must be a dirty forward price.

impliedYield

Arguments

:: GenForward f

-> Double
underlyingSpotValue

-> Double
forwarValue

-> Day
settlementDate

-> Compounding

-> DayCounter

-> IO InterestRate

Simple yield calculation based on underlying spot and forward values, taking into account underlying income. When $ t>0 $, call with: underlyingSpotValue=spotValue(t), forwardValue=strikePrice, to get current yield. For a repo, if $ t=0 $, impliedYield should reproduce the spot repo rate. For FRA's, this should reproduce the relevant zero rate at the FRA's maturityDate_;

settlementDate :: GenForward f -> IO Day

Date on which the forward contract settles.

spotIncome :: GenForward f -> GenYieldTermStructure y -> IO Double

NPV of incomedividendsstorage-costs etc. of underlying instrument.

spotValue :: GenForward f -> IO Double

returns spot value/price of an underlying financial instrument

forwardRate :: ForwardRateAgreement -> IO InterestRate

Returns the relevant forward rate associated with the FRA term.

fairForwardRate :: FxForward -> IO Double

The market-implied fair forward rate, computed by the pricing engine.

npvSourceCurrency :: FxForward -> IO Double

NPV in source currency terms.

npvTargetCurrency :: FxForward -> IO Double

NPV in target currency terms.