QuantLib.Quote

Documentation

type Quote = GenQuote CQuote

type SimpleQuote = GenQuote CSimpleQuote

data GenQuote a

asQuote :: GenQuote a -> IO Quote

data PriceType

Constructors

Bid

Ask

Last

Close

Mid

MidEquivalent

MidSafe

data IntervalPriceType

Constructors

IntervalPriceOpen

IntervalPriceClose

IntervalPriceHigh

IntervalPriceLow

data AtmType

Constructors

AtmNull

AtmSpot

AtmFwd

AtmDeltaNeutral

AtmVegaMax

AtmGammaMax

AtmPutCall50

data DeltaType

Constructors

Spot

Fwd

PaSpot

PaFwd

simpleQuote :: Double -> IO SimpleQuote

market element returning a stored value

value :: GenQuote a -> IO Double

Returns the current value of the given Quote object

isValid :: GenQuote a -> IO Bool

setValue :: SimpleQuote -> Double -> IO Double

returns the difference between the new value and the old value NB The change will propagate to all users of the quote

eurodollarFuturesImpliedStdDevQuote

Arguments

:: GenQuote a
forward

-> GenQuote b
callPrice

-> GenQuote c
putPrice

-> Double
strike

-> Double
guess

-> Double
accuracy

-> Word
maxIter

-> IO Quote

forwardSwapQuote

Arguments

:: GenSwapIndex b

-> GenQuote a
spread

-> (Int, TimeUnit)
fwdStart

-> IO Quote

forwardValueQuote :: GenIndex a -> Day -> IO Quote

futuresConvAdjustmentQuote'

Arguments

:: GenIborIndex i

-> String
immCode

-> GenQuote a
futuresQuote

-> GenQuote b
volatility

-> GenQuote c
meanReversion

-> IO Quote

futuresConvAdjustmentQuote

Arguments

:: GenIborIndex i

-> Day
futuresDate

-> GenQuote a
futuresQuote

-> GenQuote b
volatility

-> GenQuote c
meanReversion

-> IO Quote

impliedStdDevQuote

Arguments

:: OptionType

-> GenQuote a
forward

-> GenQuote b
price

-> Double

-> Double
guess

-> Double
accuracy

-> Word
maxIter

-> IO Quote

lastFixingQuote :: GenIndex a -> IO Quote

returns true if the Quote holds a valid value