QuantLib.Syntax

Documentation

free1st :: Name -> ExpQ

make a function with the first argument put at the last position, sort of any arity flip to make it easier to chain monadic calls

free2nd :: Name -> ExpQ

free1st' :: Int -> ExpQ

make a function with the first argument put at the last position with user supplied arity

free2nd' :: Int -> ExpQ

freeNth :: Int -> Name -> ExpQ

freeNth' :: Int -> Int -> ExpQ