QuantLib.TermStructure

Documentation

type TermStructure = GenTermStructure CTermStructure

TermStructure = GenTermStructure t
 YieldTermStructure = GenYieldTermStructure y = GenTermStructure t
   FittedBondDiscountCurve = GenYieldTermStructure ...
   RelinkableYieldTermStructure = GenYieldTermStructure ...

(MultiCurve, below with the other standalone leaves, is not a YieldTermStructure member -- it manages a cycle of them, handing out YieldTermStructure handles via addBootstrappedCurve / addNonBootstrappedCurve. See its own definition's comment.) > VolatilityTermStructure > OptionletVolatilityStructure > RelinkableOptionletVolatilityStructure > BlackVolTermStructure > BlackVarianceCurve > BlackVolatilitySurfaceDelta > RelinkableBlackVolTermStructure > SwaptionVolatilityStructure > RelinkableSwaptionVolatilityStructure > SabrSwaptionVolatilityCube > NoArbSabrSwaptionVolatilityCube > InterpolatedSwaptionVolatilityCube > SwaptionVolatilityMatrix > CapFloorTermVolatilityStructure* > CapFloorTermVolCurve > CapFloorTermVolSurface > BlackAtmVolCurve* > AbcdAtmVolCurve > BlackVolSurface* > SabrVolSurface > LocalVolTermStructure > GridModelLocalVolSurface > YoYOptionletVolatilitySurface > CPIVolatilitySurface > CallableBondVolatilityStructure > DefaultProbabilityTermStructure > ZeroInflationTermStructure > YoYInflationTermStructure > YoYCapFloorTermPriceSurface > CPICapFloorTermPriceSurface > CommodityCurve

data GenTermStructure t

asTermStructure :: GenTermStructure t -> IO TermStructure

referenceDate :: GenTermStructure t -> IO Day

the date at which discount = 1.0 and/or variance = 0.0

maxDate :: GenTermStructure t -> IO Day

the latest date for which the curve can return values