QuantLib.Time.Calendar

Documentation

data JointCalendarRule

Constructors

JoinHolidays

JoinBusinessDays

data CalendarConstructor

Constructors

Argentina

AustraliaSettlement

AustraliaASX

AustriaSettlement

AustriaExchange

Botswana

BrazilSettlement

BrazilExchange

CanadaSettlement

CanadaTSX

ChinaSSE

ChinaIB

CzechRepublic

Denmark

Finland

FranceSettlement

FranceExchange

GermanySettlement

GermanyFrankfurtStockExchange

GermanyXetra

GermanyEurex

GermanyEuwax

HongKong

Hungary

Iceland

India

IndonesiaBEJ

IndonesiaJSX

IndonesiaIDX

IsraelSettlement

IsraelTASE

IsraelSHIR

IsraelTelbor

ItalySettlement

ItalyExchange

Japan

Mexico

NewZealandWellington

NewZealandAuckland

Norway

Null

PolandSettlement

PolandWSE

RomaniaPublic

RomaniaBVB

RussiaSettlement

RussiaMOEX

SaudiArabia

Singapore

Slovakia

SouthAfrica

SouthKoreaSettlement

SouthKoreaKRX

Sweden

Switzerland

Taiwan

TARGET

Thailand

Turkey

Ukraine

UnitedKingdomSettlement

UnitedKingdomExchange

UnitedKingdomMetals

UnitedStatesSettlement

UnitedStatesNYSE

UnitedStatesGovernmentBond

UnitedStatesNERC

UnitedStatesLiborImpact

UnitedStatesFederalReserve

UnitedStatesSOFR

WeekendsOnly

Chile

Croatia

Malta

Montenegro

NorthMacedonia

Serbia

Slovenia

Uzbekistan

Bespoke !String ![Weekday]

Joint2 !Calendar !Calendar !JointCalendarRule

Joint3 !Calendar !Calendar !Calendar !JointCalendarRule

Joint4 !Calendar !Calendar !Calendar !Calendar !JointCalendarRule

data Calendar

calendar :: CalendarConstructor -> IO Calendar

adjust :: Calendar -> Day -> BusinessDayConvention -> IO Day

Adjusts a non-business day to the appropriate near business day with respect to the given convention

advance

Arguments

:: Calendar

-> Day

-> (Int, TimeUnit)

-> BusinessDayConvention

-> Bool
endOfMonth

-> IO Day

Advances the given date of the given number of business days and returns the result using business day convention and the EOM flag

addHoliday :: Calendar -> Day -> IO ()

Adds a date to the set of holidays for the given calendar.

businessDaysBetween

Arguments

:: Calendar

-> Day

-> Day

-> Bool
includeFirst

-> Bool
includeLast

-> IO Int

Calculates the number of business days between two given dates and returns the result.

endOfMonth :: Calendar -> Day -> IO Day

last business day of the month to which the given date belongs

isBusinessDay :: Calendar -> Day -> IO Bool

Returns true iff the date is a business day for the given market.

isEndOfMonth :: Calendar -> Day -> IO Bool

Returns true iff the date is last business day for the month in given market.

isHoliday :: Calendar -> Day -> IO Bool

Returns true iff the date is a holiday for the given market.

isWeekend :: Calendar -> Weekday -> IO Bool

Returns true iff the weekday is part of the weekend for the given market.

removeHoliday :: Calendar -> Day -> IO ()

Removes a date from the set of holidays for the given calendar.

holidays

Arguments

:: Calendar

-> Day
from

-> Day
to

-> Bool
includeWeekEnds

-> IO [Day]

Returns the holidays between two dates.

data BusinessDayConvention

Constructors

Following

ModifiedFollowing

Preceding

ModifiedPreceding

Unadjusted

HalfMonthModifiedFollowing

Nearest