QuantLib.Time.Date

Date values and fields

data Day

The Modified Julian Day is a standard count of days, with zero being the day 1858-11-17.

minDate :: Day

earliest allowed date in QuantLib

maxDate :: Day

latest date allowed in QuantLib

today :: IO Day

isLeap :: Day -> Bool

year :: Day -> Int

month :: Day -> Month

weekday :: Day -> IO Weekday

the day of the week for the given date

Calendar enums

data Month

returns TRUE if the given date's year is leap

Constructors

January

Jan

February

Feb

March

Mar

April

Apr

May

June

Jun

July

Jul

August

Aug

September

Sep

October

Oct

November

Nov

December

Dec

data Weekday

Constructors

Sunday

Sun

Monday

Mon

Tuesday

Tue

Wednesday

Wed

Thursday

Thu

Friday

Fri

Saturday

Sat

data ImmMonth

Constructors

F

G

H

J

K

M

N

Q

U

V

X

Z

Month constructors

january :: Int -> Int -> Day

helper function that is convenient for use as an infix operator

february :: Int -> Int -> Day

march :: Int -> Int -> Day

april :: Int -> Int -> Day

may :: Int -> Int -> Day

june :: Int -> Int -> Day

july :: Int -> Int -> Day

august :: Int -> Int -> Day

september :: Int -> Int -> Day

october :: Int -> Int -> Day

november :: Int -> Int -> Day

december :: Int -> Int -> Day

Date queries and arithmetic

dayOfYear :: Day -> IO Int

One-based (Jan 1st = 1)

endOfMonth :: Day -> IO Day

last day of the month to which the given date belongs

isEndOfMonth :: Day -> IO Bool

whether a date is the last day of its month

nextWeekday :: Day -> Weekday -> IO Day

next given weekday following or equal to the given date E.g., the Friday following Tuesday, January 15th, 2002 was January 18th, 2002.see http://www.cpearson.com/excel/DateTimeWS.htm

nthWeekday :: Word -> Weekday -> Month -> Int -> IO Day

n-th given weekday in the given month and year E.g., the 4th Thursday of March, 1998 was March 26th, 1998.see http://www.cpearson.com/excel/DateTimeWS.htm

IMM dates and codes

immCode :: Day -> IO String

returns the IMM code for the given date (e.g. H3 for March 20th, 2013). Warning It raises an exception if the input date is not an IMM date

immDate :: String -> Day -> IO Day

returns the IMM date for the given IMM code (e.g. March 20th, 2013 for H3). Warning It raises an exception if the input string is not an IMM code

isImmCode

Arguments

:: String
immCode

-> Bool
mainCycle

-> Bool

returns whether or not the given string is an IMM code

isImmDate

Arguments

:: Day

-> Bool
mainCycle

-> IO Bool

returns whether or not the given date is an IMM date

nextImmCode

Arguments

:: Day

-> Bool
mainCycle

-> IO String

next IMM code following the given date returns the IMM code for next contract listed in the International Money Market section of the Chicago Mercantile Exchange.

nextImmCodeFromCode

Arguments

:: String

-> Bool
mainCycle

-> Day

-> IO String

next IMM code following the given code returns the IMM code for next contract listed in the International Money Market section of the Chicago Mercantile Exchange.

nextImmDate

Arguments

:: Day

-> Bool
mainCycle

-> IO Day

next IMM date following the given date returns the 1st delivery date for next contract listed in the International Money Market section of the Chicago Mercantile Exchange.

nextImmDateFromCode

Arguments

:: String

-> Bool
mainCycle

-> Day
referenceDate

-> IO Day

next IMM date following the given IMM code returns the 1st delivery date for next contract listed in the International Money Market section of the Chicago Mercantile Exchange.

addPeriod :: Day -> (Int, TimeUnit) -> IO Day

the given date advanced by a period

ECB maintenance dates

addEcbDate :: Day -> IO ()

adds a date to the set of known ECB maintenance period start dates

ecbCode :: Day -> IO String

returns the ECB code for the given date (e.g. MAR10 for March xxth, 2010).Warning It raises an exception if the input date is not an ECB date

ecbDateFromCode :: String -> Maybe Day -> IO Day

returns the ECB date for the given ECB code (e.g. March xxth, 2013 for MAR10).WarningIt raises an exception if the input string is not an ECB code

ecbDate :: Month -> Int -> IO Day

maintenance period start date in the given month/year

isEcbCode :: String -> IO Bool

returns whether or not the given string is an ECB code

isEcbDate :: Day -> IO Bool

returns whether or not the given date is a maintenance period start date

knownEcbDates :: IO [Day]

the set of known ECB maintenance period start dates

nextEcbCodeFromCode :: String -> IO String

next ECB code following the given code

nextEcbCode :: Maybe Day -> IO String

next ECB code following the given date

nextEcbDateFromCode :: String -> Maybe Day -> IO Day

next maintenance period start date following the given ECB code

nextEcbDate :: Maybe Day -> IO Day

next maintenance period start date following the given date

nextEcbDatesFromCode :: String -> Maybe Day -> IO [Day]

next maintenance period start dates following the given code

nextEcbDates :: Maybe Day -> IO [Day]

next maintenance period start dates following the given date

removeEcbDate :: Day -> IO ()

removes a date from the set of known ECB maintenance period start dates