QuantLib.Index

Documentation

type Index = GenIndex CIndex

data GenIndex a

addFixing

Arguments

:: GenIndex a

-> Day

-> Double

-> Bool
forceOverwrite

-> IO ()

stores the historical fixing at the given date the date passed as arguments must be the actual calendar date of the fixing; no settlement days must be used. Adds fixings for the given InterestRateIndex object

returns the calendar defining valid fixing dates

fixingCalendar :: GenIndex a -> IO Calendar

asIndex :: GenIndex a -> IO Index

Orphan instances

Show Index

Instance details

Methods

showsPrec :: Int -> Index -> ShowS

show :: Index -> String

showList :: [Index] -> ShowS