QuantLib.Instrument.Swap

Documentation

type Swaption = GenOption CSwaption

type Swap = GenSwap CSwap

type VanillaSwap = GenSwap CVanillaSwap

type AssetSwap = GenSwap CAssetSwap

type OvernightIndexedSwap = GenSwap COvernightIndexedSwap

type BMASwap = GenSwap CBMASwap

asSwap :: GenSwap a -> IO Swap

impliedVolatility

Arguments

:: Swaption

-> Double
price

-> GenYieldTermStructure y

-> Double
guess

-> Double
accuracy

-> Word
maxEvaluations

-> Double
minVol

-> Double
maxVol

-> IO Double

implied volatility

data SwapType

Constructors

Receiver

Payer

swap'

Arguments

:: [(Leg, Bool)]
(legs, payer)

-> IO Swap

Multi leg constructor.

swap :: GenLeg a -> GenLeg a -> IO Swap

The cash flows belonging to the first leg are paid; the ones belonging to the second leg are received.

bmaSwap

Arguments

:: SwapType

-> Double
nominal

-> Schedule
liborSchedule

-> Double
liborFraction

-> Double
liborSpread

-> GenIborIndex a

-> DayCounter
liborDayCount

-> Schedule
bmaSchedule

-> BMAIndex

-> DayCounter
bmaDayCount

-> IO BMASwap

vanillaSwap

Arguments

:: SwapType

-> Double
nominal

-> Schedule
fixedSchedule

-> Double
fixedRate

-> DayCounter
fixedDayCount

-> Schedule
floatSchedule

-> GenIborIndex a

-> Double
spread

-> DayCounter
floatingDayCount

-> BusinessDayConvention
paymentConvention

-> IO VanillaSwap

endDiscounts :: GenSwap a -> Word -> IO Double

leg :: GenSwap a -> Word -> IO Leg

legBPS :: GenSwap a -> Word -> IO Double

legNPV :: GenSwap a -> Word -> IO Double

maturityDate :: GenSwap a -> IO (Maybe Day)

npvDateDiscount :: GenSwap a -> IO Double

startDate :: GenSwap a -> IO (Maybe Day)

startDiscounts :: GenSwap a -> Word -> IO Double

bmaLeg :: BMASwap -> IO Leg

bmaLegBPS :: BMASwap -> IO Double

bmaLegNPV :: BMASwap -> IO Double

fairLiborFraction :: BMASwap -> IO Double

fairLiborSpread :: BMASwap -> IO Double

liborFraction :: BMASwap -> IO Double

liborLeg :: BMASwap -> IO Leg

liborLegBPS :: BMASwap -> IO Double

liborLegNPV :: BMASwap -> IO Double

swaption :: VanillaSwap -> Exercise -> SettlementType -> IO Swaption

assetSwap

Arguments

:: Bool
payBondCoupon

-> Bond

-> Double
bondCleanPrice

-> GenIborIndex a

-> Double

-> Schedule
floatSchedule

-> DayCounter
floatingDayCount

-> Bool
parAssetSwap

-> IO AssetSwap

bondLeg :: AssetSwap -> IO Leg

cleanPrice :: AssetSwap -> IO Double

fairCleanPrice :: AssetSwap -> IO Double

fairNonParRepayment :: AssetSwap -> IO Double

nonParRepayment :: AssetSwap -> IO Double

parSwap :: AssetSwap -> IO Bool

payBondCoupon :: AssetSwap -> IO Bool

overnightIndexedSwap

Arguments

:: SwapType

-> Double
nominal

-> Schedule

-> Double
fixedRate

-> DayCounter
fixedDC

-> OvernightIborIndex

-> Double
spread

-> IO OvernightIndexedSwap

overnightIndexedSwap'

Arguments

:: SwapType

-> [Double]
nominals

-> Schedule
schedule

-> Double
fixedRate

-> DayCounter
fixedDC

-> OvernightIborIndex

-> Double
spread

-> IO OvernightIndexedSwap

overnightLeg :: OvernightIndexedSwap -> IO Leg

overnightLegBPS :: OvernightIndexedSwap -> IO Double

overnightLegNPV :: OvernightIndexedSwap -> IO Double

class HasFixedLeg a where

Methods

fairRate :: a -> IO Double

fixedLeg :: a -> IO Leg

fixedLegBPS :: a -> IO Double

fixedLegNPV :: a -> IO Double

class HasFloatingLeg a where

Methods

floatingLeg :: a -> IO Leg

floatingLegBPS :: a -> IO Double

floatingLegNPV :: a -> IO Double

class HasSpread a where

Methods

fairSpread :: a -> IO Double