QuantLib.InterestRate

Documentation

data Compounding

Constructors

Simple

Compounded

Continuous

SimpleThenCompounded

CompoundedThenSimple

data VolatilityType

Constructors

ShiftedLognormal

Normal

data InterestRate

interestRate

Arguments

:: Double
r

-> DayCounter

-> Compounding

-> Frequency

-> IO InterestRate

construct an interest rate from a rate value, a day counter, a compounding convention and a frequency.

compoundFactor

Arguments

:: InterestRate

-> Double
t

-> IO Double

compound factor implied by the rate compounded at time t. returns the compound (a.k.a capitalization) factor implied by the rate compounded at time t. Warning Time must be measured using InterestRate's own day counter.

compoundFactor'

Arguments

:: InterestRate

-> Day
d1

-> Day
d2

-> Day
refStart

-> Day
refEnd

-> IO Double

compound factor implied by the rate compounded between two dates returns the compound (a.k.a capitalization) factor implied by the rate compounded between two dates.

discountFactor :: InterestRate -> Double -> IO Double

discount factor implied by the rate compounded at time t. Warning Time must be measured using InterestRate's own day counter.

discountFactor'

Arguments

:: InterestRate

-> Day
d1

-> Day
d2

-> Day
refStart

-> Day
refEnd

-> IO Double

discount factor implied by the rate compounded between two dates

equivalentRate

Arguments

:: InterestRate

-> Compounding

-> Frequency

-> Double
t

-> IO InterestRate

equivalent interest rate for a compounding period t. The resulting InterestRate shares the same implicit day-counting rule of the original InterestRate instance. Warning Time must be measured using the InterestRate's own day counter.

equivalentRate'

Arguments

:: InterestRate

-> DayCounter
resultDC

-> Compounding

-> Frequency

-> Day
d1

-> Day
d2

-> Day
refStart

-> Day
refEnd

-> IO InterestRate

equivalent rate for a compounding period between two dates The resulting rate is calculated taking the required day-counting rule into account.

impliedRate

Arguments

:: InterestRate

-> Double
compound

-> DayCounter

-> Compounding

-> Frequency

-> Double
t

-> IO InterestRate

implied interest rate for a given compound factor at a given time. The resulting InterestRate has the day-counter provided as input. Warning Time must be measured using the day-counter provided as input.

impliedRate'

Arguments

:: InterestRate

-> Double
compound

-> DayCounter

-> Compounding

-> Frequency

-> Day
d1

-> Day
d2

-> Day
refStart

-> Day
refEnd

-> IO InterestRate

implied rate for a given compound factor between two dates. The resulting rate is calculated taking the required day-counting rule into account.

rate :: InterestRate -> Double

the rate value of an interest rate.