QuantLib.Math

Documentation

data RoundingType

Constructors

None

Up

Down

Closest

Floor

Ceiling

data Rounding

Constructors

NoRounding

Rounding

Fields

!Int

precision

!RoundingType

!Int

digit

applyRounding :: Rounding -> Double -> Double

data EndCriteriaType

Constructors

EndNone

MaxIterations

StationaryPoint

StationaryFunctionValue

StationaryFunctionAccuracy

ZeroGradientNorm

Unknown

data HistogramAlgorithm

Constructors

HistogramNone

Sturges

FD

Scott

data Approximation

Constructors

NaturalSpline !Bool

Parabolic !Bool

Kruger

FritschButland

data Interpolation

Constructors

BackwardFlat

ForwardFlat

Linear

LogLinear

Cubic !Approximation

LogCubic !Approximation

Abcd

data RngTrait

Constructors

PseudoRandom

PoissonPseudoRandom

LowDiscrepancy

Ziggurat

data BinomialTree

Constructors

JarrowRudd

CoxRossRubinstein

AdditiveEQPBinomialTree

Trigeorgis

Tian

LeisenReimer

Joshi4

ExtendedJarrowRudd

ExtendedCoxRossRubinstein

ExtendedAdditiveEQPBinomialTree

ExtendedTrigeorgis

ExtendedTian

ExtendedLeisenReimer

ExtendedJoshi4

data BoundaryConditionSide

Constructors

BoundaryNone

Upper

Lower

data FdmSchemeType

Constructors

HundsdorferType

DouglasType

CraigSneydType

ModifiedCraigSneydType

ImplicitEulerType

ExplicitEulerType

MethodOfLinesType

TrBDF2Type

CrankNicolsonType

data FdmScheme

Constructors

FdmScheme

Fields

!FdmSchemeType

type

!Double

theta

!Double

mu

CraigSneyd

Douglas

ExplicitEuler

Hundsdorfer

ImplicitEuler

ModifiedCraigSneyd

ModifiedHundsdorfer

data PolynomialType

Constructors

Monomial

Laguerre

Hermite

Hyperbolic

Legendre

Chebyshev

Chebyshev2nd

data ComplexLogFormula

Constructors

Gatheral

BranchCorrection

AndersenPiterbarg

AndersenPiterbargOptCV

AsymptoticChF

OptimalCV

data CmsMarketCalibrationType

Constructors

OnSpread

OnPrice

OnForwardCmsPrice

data EndCriteria

Constructors

EndCriteria

Fields

!Word

maxIterations

!Word

maxStationaryStateIterations

!Double

rootEpsilon

!Double

functionEpsilon

!Double

gradientNormEpsilon

data OptimizationMethod

Constructors

LevenbergMarquardt

Fields

!Double

epsfcn

!Double

xtol

!Double

gtol

!Bool

useCostFunctionsJacobian

Simplex !Double
lambda, characteristic length

data Constraint

Constructors

Boundary

Fields

!Double

low

!Double

high

Composite

Fields

!Constraint

c1

!Constraint

c2

NoConstraint

PositiveConstraint

data SobolDirectionIntegers

Constructors

Unit

Jaeckel

SobolLevitan

SobolLevitanLemieux

JoeKuoD5

JoeKuoD6

JoeKuoD7

Kuo

Kuo2

Kuo3

data Matrix a

realMatrix :: Word -> Word -> [Double] -> Either String (Matrix Double)

objectMatrix :: Word -> Word -> [a] -> Either String (Matrix a)

data TimeGrid

timeGrid

Arguments

:: Double
end

-> Word
steps

-> IO TimeGrid

Regularly spaced time-grid.

timeGridFromList :: NonEmpty Double -> IO TimeGrid

Time grid with mandatory time points. Mandatory points are guaranteed to belong to the grid. No additional points are added.

timeGridFromList' :: NonEmpty Double -> Word -> IO TimeGrid

Time grid with mandatory time points. Mandatory points are guaranteed to belong to the grid. Additional points are then added with regular spacing between pairs of mandatory times in order to reach the desired number of steps.

timeAt :: TimeGrid -> Word -> IO Double

size :: TimeGrid -> Word

points :: TimeGrid -> IO [Double]

points' :: TimeGrid -> IO (Vector CDouble)