QuantLib.Settings

Description

global repository for run-time library settings

Documentation

evaluationDate :: IO Day

returns the current value of the boolean which enforce the usage of historic fixings for today's date

setEvaluationDate :: Maybe Day -> IO ()

sets the value of the Evaluation Date |Nothing sets the evaluation date to Date::todaysDate() and allow it to change at midnight. This comes at the price of losing some performance, since the evaluation date is re-evaluated each time it is read.

returns the current value of the Evaluation Date: the date at which pricing is to be performed

enforceTodaysHistoricFixings :: IO Bool

setEnforceTodaysHistoricFixings :: Bool -> IO ()

sets the value of the boolean which enforce the usage of historic fixings for today's date

includeTodaysCashFlows :: IO (Maybe Bool)

if set, whether CashFlows occurring on today's date should enter the NPV; when the NPV date equals today's date this overrides includeReferenceDateEvents and cannot be overridden locally

setIncludeTodaysCashFlows :: Maybe Bool -> IO ()

sets whether CashFlows occurring on today's date should enter the NPV

includeReferenceDateEvents :: IO Bool

This flag specifies whether or not Events occurring on the reference date should, by default, be taken into account as not happened yet. It can be overridden locally when calling the Event::hasOccurred method.

setIncludeReferenceDateEvents :: Bool -> IO ()

sets whether Events occurring on the reference date should, by default, be taken into account as not happened yet

keepingSettings :: IO b -> IO b

brackets to restore settings once action has completed or raised an exception

keepingSettings' :: IO b -> IO b

brackets to restore settings once action has completed or raised an exception. Before restoring settings garbage collection is run to avoid problems with market data objects watching evaluation date

version :: String

boostVersion :: String

epsilon :: Double