QuantLib.TermStructure.Credit

Documentation

data ProbabilityTrait

Constructors

SurvivalProbability

HazardRate

DefaultDensity

type DefaultProbabilityTermStructure = GenTermStructure CDefaultProbabilityTermStructure

data DefaultProbabilityHelper

factorSpreadedHazardRateCurve :: DefaultProbabilityTermStructure -> GenQuote q -> IO DefaultProbabilityTermStructure

flatHazardRate' :: Word -> Calendar -> GenQuote q -> DayCounter -> IO DefaultProbabilityTermStructure

flat hazard-rate curve anchored at a settlement date

flatHazardRate :: Day -> GenQuote q -> DayCounter -> IO DefaultProbabilityTermStructure

flat hazard-rate curve anchored at a reference date

spreadedHazardRateCurve :: DefaultProbabilityTermStructure -> GenQuote q -> IO DefaultProbabilityTermStructure

a curve whose survival probability is another curve's, multiplied by a spread factor

defaultProbability

Arguments

:: DefaultProbabilityTermStructure

-> Day

-> Bool
extrapolate

-> IO Double

default probability from the reference date until a given date

hazardRate'

Arguments

:: DefaultProbabilityTermStructure

-> Double

-> Bool
extrapolate

-> IO Double

hazard rate at a given time, with annual frequency and continuous compounding

hazardRate

Arguments

:: DefaultProbabilityTermStructure

-> Day

-> Bool
extrapolate

-> IO Double

hazard rate at a given date, with annual frequency and continuous compounding

survivalProbability'

Arguments

:: DefaultProbabilityTermStructure

-> Double

-> Bool
extrapolate

-> IO Double

The same day-counting rule used by the term structure should be used for calculating the passed time t.

survivalProbability

Arguments

:: DefaultProbabilityTermStructure

-> Day

-> Bool
extrapolate

-> IO Double

survival probability from the reference date until a given date

defaultDensity'

Arguments

:: DefaultProbabilityTermStructure

-> Double

-> Bool
extrapolate

-> IO Double

The same day-counting rule used by the term structure should be used for calculating the passed time t.

defaultDensity

Arguments

:: DefaultProbabilityTermStructure

-> Day

-> Bool
extrapolate

-> IO Double

default density at a given date

defaultProbability'

Arguments

:: DefaultProbabilityTermStructure

-> Double

-> Bool
extrapolate

-> IO Double

The same day-counting rule used by the term structure should be used for calculating the passed time t.

defaultProbabilityBetween

Arguments

:: DefaultProbabilityTermStructure

-> Day

-> Day

-> Bool
extrapolate

-> IO Double

probability of default between two given dates

defaultProbabilityBetween'

Arguments

:: DefaultProbabilityTermStructure

-> Double

-> Double

-> Bool
extrapolate

-> IO Double

probability of default between two given times

spreadCdsHelper

Arguments

:: GenQuote q
runningSpread

-> (Word, TimeUnit)
tenor

-> Int
settlementDays

-> Calendar

-> Frequency

-> BusinessDayConvention

-> DateGenerationRule

-> DayCounter

-> Double

-> GenYieldTermStructure y
discountCurve

-> Bool
settlesAccrual

-> Bool
paysAtDefaultTime

-> Maybe Day
startDate

-> DayCounter
lastPeriodDayCounter

-> Bool
rebatesAccrual

-> PricingModel
model

-> IO DefaultProbabilityHelper

bootstrap helper for a CDS quoted by running spread

upfrontCdsHelper

Arguments

:: GenQuote q
upfront

-> Double
runningSpread

-> (Word, TimeUnit)
tenor

-> Int
settlementDays

-> Calendar

-> Frequency

-> BusinessDayConvention

-> DateGenerationRule

-> DayCounter

-> Double
recoveryDate

-> GenYieldTermStructure y
discountCurve

-> Word
upfrontSettlementDays

-> Bool

-> Bool
paysAtDefaultTime

-> Maybe Day
startDate

-> DayCounter
lastPeriodDayCounter

-> Bool
rebatesAccrual

-> PricingModel
model

-> IO DefaultProbabilityHelper

the upfront must be quoted in fractional units.

interpolatedDefaultDensityCurve

Arguments

:: [(Day, Double)]

-> DayCounter

-> Calendar

-> [(Day, GenQuote q)]
jumps

-> Interpolation

-> IO DefaultProbabilityTermStructure

interpolatedHazardRateCurve

Arguments

:: [(Day, Double)]

-> DayCounter

-> Calendar

-> [(Day, GenQuote q)]
jumps

-> Interpolation

-> Bool
extrapolate past the curve's max date

-> IO DefaultProbabilityTermStructure

interpolatedSurvivalProbabilityCurve

Arguments

:: [(Day, Double)]

-> DayCounter

-> Calendar

-> [(Day, GenQuote q)]
jumps

-> Interpolation

-> IO DefaultProbabilityTermStructure

piecewiseDefaultCurve

Arguments

:: Day

-> [DefaultProbabilityHelper]

-> DayCounter

-> [(Day, GenQuote q)]
jumps

-> ProbabilityTrait

-> Interpolation

-> IO DefaultProbabilityTermStructure

piecewiseDefaultCurve'

Arguments

:: Word

-> Calendar

-> [DefaultProbabilityHelper]

-> DayCounter

-> [(Day, GenQuote q)]
jumps

-> ProbabilityTrait

-> Interpolation

-> IO DefaultProbabilityTermStructure