QuantLib.TermStructure.Yield

Documentation

type YieldTermStructure = GenYieldTermStructure CYieldTermStructure

type GenYieldTermStructure a = GenTermStructure (AnyOf CYieldTermStructure' a)

type BondHelper = GenRateHelper CBondHelper

type RateHelper = GenRateHelper CRateHelper

type SwapRateHelper = GenRateHelper CSwapRateHelper

type OISRateHelper = GenRateHelper COISRateHelper

data FittingMethod

Constructors

CubicBSplines

Fields

![Double]

knotVector (year fraction)

!Bool

constrainAtZero

ExponentialSplines !Bool

NelsonSiegel

SimplePolynomial

Fields

!Word

degree

!Bool

constrainAtZero

Svensson

type FittedBondDiscountCurve = GenYieldTermStructure CFittedBondDiscountCurve

fittedBondDiscountCurve

Arguments

:: Word
settlementDays

-> Calendar

-> [BondHelper]

-> DayCounter

-> FittingMethod

-> Double
accuracy

-> Word
maxEvaluations

-> [Double]
guess

-> Double
simplexLambda

-> IO FittedBondDiscountCurve

fittedBondDiscountCurve'

Arguments

:: Day

-> [BondHelper]

-> DayCounter

-> FittingMethod

-> Double
accuracy

-> Word
maxEvaluations

-> [Double]
guess

-> Double
simplexLambda

-> IO FittedBondDiscountCurve

curve reference date fixed for life of curve

data GenRateHelper a

data BootstrapTrait

Constructors

Discount

ZeroYield

ForwardRate

depositRateHelper' :: GenQuote a -> GenIborIndex b -> IO RateHelper

depositRateHelper

Arguments

:: GenQuote a
rate

-> (Int, TimeUnit)
tenor

-> Word
fixingDays

-> Calendar
calendar

-> BusinessDayConvention
convention

-> Bool
endOfMonth

-> DayCounter

-> IO RateHelper

fixedRateBondHelper :: GenQuote a -> Word -> Double -> Schedule -> [Double] -> DayCounter -> BusinessDayConvention -> Double -> Maybe Day -> IO BondHelper

discount'

Arguments

:: GenYieldTermStructure a

-> Day
d

-> Bool
extrapolate

-> IO Double

Returns a discount factor from the given YieldTermStructure object

swapRateHelper'

Arguments

:: GenQuote a
rate

-> (Int, TimeUnit)
tenor

-> Calendar
calendar

-> Frequency
fixedFrequency

-> BusinessDayConvention
fixedConvention

-> DayCounter
fixedDayCount

-> GenIborIndex b
iborIndex

-> Maybe (GenQuote s)
spread

-> (Int, TimeUnit)
fwdStart

-> Maybe (GenYieldTermStructure d)
discountingCurve

-> IO SwapRateHelper

flatForward :: Day -> GenQuote a -> DayCounter -> Compounding -> Frequency -> IO YieldTermStructure

flatForward'

Arguments

:: Word
settlementDays

-> Calendar

-> GenQuote a

-> DayCounter

-> Compounding

-> Frequency

-> IO YieldTermStructure

The resulting interest rate has the required daycounting rule.

zeroRate'

Arguments

:: GenYieldTermStructure a

-> Day

-> DayCounter

-> Compounding

-> Frequency

-> Bool
extrapolate

-> IO InterestRate

forwardRateForPeriod

Arguments

:: GenYieldTermStructure a

-> Day

-> (Int, TimeUnit)

-> DayCounter

-> Compounding

-> Frequency

-> Bool
extrapolate

-> IO InterestRate

The resulting interest rate has the required day-counting rule. Warning dates are not adjusted for holidays

forwardRate'

Arguments

:: GenYieldTermStructure a

-> Day

-> Day

-> DayCounter

-> Compounding

-> Frequency

-> Bool
extrapolate

-> IO InterestRate

The resulting interest rate has the required day-counting rule.

forwardRate

Arguments

:: GenYieldTermStructure a

-> Double

-> Double

-> Compounding

-> Frequency

-> Bool
extrapolate

-> IO InterestRate

The resulting interest rate has the same day-counting rule used by the term structure. The same rule should be used for calculating the passed times t1 and t2.

zeroRate

Arguments

:: GenYieldTermStructure a

-> Double

-> Compounding

-> Frequency

-> Bool
extrapolate

-> IO InterestRate

The resulting interest rate has the same day-counting rule used by the term structure. The same rule should be used for calculating the passed time t.

The same day-counting rule used by the term structure should be used for calculating the passed time t.

discount

Arguments

:: GenYieldTermStructure a

-> Double

-> Bool
extrapolate

-> IO Double

fraRateHelper

Arguments

:: GenQuote a
rate

-> Word
monthsToStart

-> Word
monthsToEnd

-> Word
fixingDays

-> Calendar
calendar

-> BusinessDayConvention
convention

-> Bool
endOfMonth

-> DayCounter

-> IO RateHelper

bondHelper :: GenQuote a -> Bond -> IO BondHelper

Warning Setting a pricing engine to the passed bond from external code will cause the bootstrap to fail or to give wrong results. It is advised to discard the bond after creating the helper, so that the helper has sole ownership of it.

oisRateHelper :: Word -> (Int, TimeUnit) -> GenQuote a -> OvernightIborIndex -> Maybe (GenYieldTermStructure b) -> IO OISRateHelper

oisRateHelper' :: Day -> Day -> GenQuote a -> OvernightIborIndex -> Maybe (GenYieldTermStructure b) -> IO OISRateHelper

swapRateHelper :: GenQuote a -> GenSwapIndex b -> Maybe (GenQuote m) -> (Int, TimeUnit) -> Maybe (GenYieldTermStructure c) -> IO SwapRateHelper

forwardSpreadedTermStructure :: GenYieldTermStructure b -> GenQuote a -> IO YieldTermStructure

zeroSpreadedTermStructure :: GenYieldTermStructure b -> GenQuote a -> Compounding -> Frequency -> IO YieldTermStructure

bmaSwapRateHelper

Arguments

:: GenQuote a
liborFraction

-> (Int, TimeUnit)
tenor

-> Word
settlementDAys

-> Calendar

-> (Int, TimeUnit)
bmpPeriod

-> BusinessDayConvention

-> DayCounter

-> BMAIndex

-> GenIborIndex b

-> IO RateHelper

fraIborRateHelper'

Arguments

:: GenQuote a

-> Word
monthsToStart

-> GenIborIndex b

-> IO RateHelper

fraRateHelper'

Arguments

:: GenQuote a

-> (Int, TimeUnit)
periodToStart

-> Word
lengthInMonths

-> Word
fixingDays

-> Calendar

-> BusinessDayConvention

-> Bool
endOfMonth

-> DayCounter

-> IO RateHelper

fraIborRateHelper

Arguments

:: GenQuote a

-> (Int, TimeUnit)
periodToStart

-> GenIborIndex b

-> IO RateHelper

futuresRateHelper'

Arguments

:: GenQuote a

-> Day
immStartDate

-> Day
endDate

-> DayCounter

-> Maybe (GenQuote m)
convexityAdjustment

-> IO RateHelper

futuresIborRateHelper

Arguments

:: GenQuote a

-> Day
immDate

-> GenIborIndex b

-> Maybe (GenQuote m)

-> IO RateHelper

futuresRateHelper

Arguments

:: GenQuote a

-> Day
immDate

-> Word
lengthInMonths

-> Calendar

-> BusinessDayConvention

-> Bool
endOfMonth

-> DayCounter

-> Maybe (GenQuote m)
convexityAdjustment

-> IO RateHelper

impliedQuote :: GenRateHelper a -> IO Double

impliedTermStructure :: GenYieldTermStructure a -> Day -> IO YieldTermStructure

asYieldTermStructure :: GenYieldTermStructure a -> IO YieldTermStructure

asRateHelper :: GenRateHelper a -> IO RateHelper

piecewiseZeroSpreadedTermStructure

Arguments

:: GenYieldTermStructure b

-> [(Day, GenQuote a)]
spreads

-> Compounding

-> Frequency

-> IO YieldTermStructure

quantoTermStructure

Arguments

:: GenYieldTermStructure a
underlyingDividendTS

-> GenYieldTermStructure b
riskFreeTS

-> GenYieldTermStructure c
foreignRsikFreeTS

-> GenBlackVolTermStructure d
underlyingBlackVolTS

-> Double
strike

-> GenBlackVolTermStructure e
exchRateBlackVolTS

-> Double
exchRateATMlevel

-> Double
underlyingExchRateCorrelation

-> IO YieldTermStructure

minimumCostValue :: FittedBondDiscountCurve -> IO Double

final value of cost function after optimization

final number of iterations used in the optimization problem

numberOfIterations :: FittedBondDiscountCurve -> IO Int

piecewiseYieldCurve

Arguments

:: Day
referenceDate

-> [GenRateHelper b]
instruments

-> DayCounter
dayCounter

-> [(Day, GenQuote a)]
jumps

-> BootstrapTrait
bootstrap trait

-> Interpolation
interpolator

-> IO YieldTermStructure

piecewiseYieldCurve'

Arguments

:: Word
settlementDays

-> Calendar
calendar

-> [GenRateHelper b]
instruments

-> DayCounter
dayCounter

-> [(Day, GenQuote a)]
jumps

-> BootstrapTrait
bootstrap trait

-> Interpolation
interpolator

-> IO YieldTermStructure

interpolatedZeroCurve

Arguments

:: [(Day, Double)]
dates, yields

-> DayCounter
dayCounter

-> Calendar
cal

-> [(Day, GenQuote a)]
jumps, jumpDates

-> Interpolation
interpolator

-> IO YieldTermStructure

interpolatedForwardCurve

Arguments

:: [(Day, Double)]
dates, forwards

-> DayCounter
dayCounter

-> Calendar
cal

-> [(Day, GenQuote a)]
jumps

-> Interpolation
interpolator

-> IO YieldTermStructure

interpolatedDiscountCurve

Arguments

:: [(Day, Double)]
dates, dfs

-> DayCounter
dayCounter

-> Calendar
cal

-> [(Day, GenQuote a)]
jumps

-> Interpolation
interpolator

-> IO YieldTermStructure

underlying :: HelperUnderlying a b => a -> IO b